fix : split Request and Request Error
