+++ src/Request.php
@@ -25,7 +25,6 @@
 
 namespace Alma\API;
 
-
 // In older versions of PHP (<= 5.5.19), those constants aren't defined – we do need them though
 if (!defined('CURL_SSLVERSION_TLSv1_0')) {
     define('CURL_SSLVERSION_TLSv1_0', 4);
@@ -37,27 +36,6 @@ if (!defined('CURL_SSLVERSION_TLSv1_1')) {
 
 if (!defined('CURL_SSLVERSION_TLSv1_2')) {
     define('CURL_SSLVERSION_TLSv1_2', 6);
-}
-
-// phpcs:disable PSR1.Classes.ClassDeclaration.MultipleClasses
-class RequestError extends \Exception
-{
-    /**
-     * @var Request|null
-     */
-    public $request;
-    /**
-     * @var Response|null
-     */
-    public $response;
-
-    public function __construct($message = "", $request = null, $response = null)
-    {
-        parent::__construct($message);
-
-        $this->request = $request;
-        $this->response = $response;
-    }
 }
 
 class Request
